\name{plot}

\alias{plot}
\alias{plot,Phenotypes,flowFrame}
\alias{plot,Phenotypes,flowFrame-method}
\alias{plot,Phenotypes,numeric}
\alias{plot,Phenotypes,numeric-method}
\alias{plot,Phenotypes,character}
\alias{plot,Phenotypes,character-method}

\docType{methods}

\title{Methods for Function plot}

\description{
Methods for function plot
}

\usage{
\S4method{plot}{Phenotypes,flowFrame}(x, y, ...)
\S4method{plot}{Phenotypes,numeric}(x, y, Frame,...)
\S4method{plot}{Phenotypes,character}(x, y, Frame,...)
}

\arguments{
\item{x}{An object of class Phenotypes as generated by the flowType package.}
\item{y}{A flowFrame or a numeric/character value representing the
  phenotype that needs to be plotted depending on the signature of the
  function}
\item{Frame}{A flowFrame (might be optional depending on the signature
  of the function}
\item{...}{Extra parameters that will be passed to the generic plot function}
}


\author{
Nima Aghaeepour <\email{naghaeep@gmail.com}>
}
\examples{
#See the vigentte
}

\seealso{
\code{\link{flowType}}
}

\keyword{print}
